pyproject.toml
[tool.poetry]
name = "test-task-bugu"
version = "0.1.0"
description = ""
authors = ["eddy.di <eddy.di.fint@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
pytest-django = "^4.8.0"
pytest-factoryboy = "^2.7.0"
faker = "^26.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
